Created scanner functions, improved Channel system
Former-commit-id: 32ad9f125290c450746d61f01857ba9fb8857485

+++ fingerprintscanner/src/main/java/com/simprints/fingerprintscanner/v2/domain/packet/Packet.kt
@@ -4,7 +4,7 @@ class Packet(
     val bytes: ByteArray,
     val header: ByteArray,
     val payload: ByteArray,
-    val source: Int,
-    val destination: Int,
+    val source: Byte,
+    val destination: Byte,
     val payloadLength: Int
 )
